tests/wpt/tests/focus/focus-restoration-in-same-site-iframes-window.html

<!doctype html>
<meta charset=utf-8>
<title>Test focus restoration</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "outerlog:log:willfocuswindow,windowfocus,didfocuswindow,windowblur,windowfocus,", 'Check log');
    }, "Check result");
    w.close();
    done();
};
var w = window.open("support/focus-restoration-in-same-site-iframes-outer-window.html");
</script>
